seecure-backend/requirements.txt
fastapi
uvicorn
python-multipart
vosk
pydantic
requests
python-dotenv
aiofiles
